add g++ compile command

+++ python/hidet/backend/build.py
@@ -52,7 +52,107 @@ def nvcc_path() -> str:
     raise FileNotFoundError('Can not find nvcc compiler.')
 
 
-def compile_source(src_path: str, out_lib_path: str, keep_ptx=False) -> None:
+@functools.lru_cache()
+def gpp_path() -> str:
+    path: Optional[str] = shutil.which('g++')
+    if path is not None:
+        return path
+
+    try_paths = ['/usr/bin/g++']
+    for path in try_paths:
+        if os.path.exists(path):
+            return path
+    raise FileNotFoundError('Can not find g++ compiler.')
+
+
+@functools.lru_cache()
+def compile_command(src_path: str, out_lib_path: str, target='cuda', keep_ptx=False):
+    # pylint: disable=too-many-locals
+
+    # dir contains the runtime header file 'hidet/runtime.h'
+    include_dirs = get_include_dirs()
+    # dir contains the runtime library 'libhidet_runtime.so'
+    library_dirs = [os.path.dirname(library_paths['hidet_runtime'])]
+    if target == 'cuda':
+        cc = hidet.cuda.compute_capability()
+        cc_code = '{}{}'.format(cc[0], cc[1])
+        # The following command compiles the cuda source code to a shared library
+        # See https://docs.nvidia.com/cuda/cuda-compiler-driver-nvcc/index.html for more information about nvcc compilation.
+        command = [
+            # the path to nvcc compiler
+            nvcc_path(),
+            # the included directories.
+            *['-I{}'.format(include_dir) for include_dir in include_dirs],
+            # the library directories.
+            *['-L{}'.format(library_dir) for library_dir in library_dirs],
+            # keep option will keep the intermediate results during compilation, including PTX.
+            '-keep' if keep_ptx else '',
+            # the target PTX and SASS version.
+            '-gencode',
+            f'arch=compute_{cc_code},code=sm_{cc_code}',
+            # allow ptxas (PTX assembler) to output information like register/smem usage.
+            '--ptxas-options=-v',
+            # compile into position independent code.
+            '--compiler-options',
+            "'-fPIC'",
+            # embed the line information into the binary, allow Nsight Compute to get the source code for profiling.
+            '-lineinfo',
+            # link the hidet runtime, all APIs for communication between kernels and host system are in hidet runtime.
+            '-lhidet_runtime',
+            # shared cuda runtime library is used (.so), instead of static one (.a). used to reduce binary size.
+            '--cudart',
+            'shared',
+            # supress some warnings
+            # see https://docs.nvidia.com/cuda/cuda-compiler-driver-nvcc/index.html#generic-tool-options-diag-suppress
+            # supress warming no 177 like: "warning #177-D: variable "xxx" was declared but never referenced"
+            '--diag-suppress 177',
+            # supress warning no 179 like: "warning #179-D: right operand of "%" is zero"
+            '--diag-suppress 179',
+            # supress warning no 39 like: "warning #39-D: division by zero"
+            '--diag-suppress 39',
+            # generate shared library (lib.so).
+            '--shared',
+            # the source path.
+            src_path,
+            # the output library path.
+            '-o',
+            out_lib_path,
+        ]
+        return command
+    else:
+        command = [
+            # the path to nvcc compiler
+            gpp_path(),
+            # the included directories.
+            *['-I{}'.format(include_dir) for include_dir in include_dirs],
+            # the library directories.
+            *['-L{}'.format(library_dir) for library_dir in library_dirs],
+            # compile into position independent code.
+            '-fPIC',
+            # embed the debug information into the binary, allow Nsight Compute to get the source code for profiling.
+            '-g',
+            # link the hidet runtime, all APIs for communication between kernels and host system are in hidet runtime.
+            '-lhidet_runtime',
+            # 'shared',
+            # generate shared library (lib.so).
+            '-shared',
+            # the source path.
+            src_path,
+            # the output library path.
+            '-o',
+            out_lib_path,
+        ]
+        return command
+
+
+def log_name(target: str) -> str:
+    if target == 'cuda':
+        return 'nvcc_log'
+    else:
+        return 'gpp_log'
+
+
+def compile_source(src_path: str, out_lib_path: str, target: str, keep_ptx=False) -> None:
     """
     Compile the source code in 'src_path' file and output the library to 'out_lib_path'.
 
@@ -68,56 +168,7 @@ def compile_source(src_path: str, out_lib_path: str, keep_ptx=False) -> None:
     # pylint: disable=too-many-locals
     src_path = os.path.abspath(src_path)
     out_lib_path = os.path.abspath(out_lib_path)
-    cc = hidet.cuda.compute_capability()
-
-    # dir contains the runtime header file 'hidet/runtime.h'
-    include_dirs = get_include_dirs()
-    # dir contains the runtime library 'libhidet_runtime.so'
-    library_dirs = [os.path.dirname(library_paths['hidet_runtime'])]
-
-    cc_code = '{}{}'.format(cc[0], cc[1])
-    # The following command compiles the cuda source code to a shared library
-    # See https://docs.nvidia.com/cuda/cuda-compiler-driver-nvcc/index.html for more information about nvcc compilation.
-    command = [
-        # the path to nvcc compiler
-        nvcc_path(),
-        # the included directories.
-        *['-I{}'.format(include_dir) for include_dir in include_dirs],
-        # the library directories.
-        *['-L{}'.format(library_dir) for library_dir in library_dirs],
-        # keep option will keep the intermediate results during compilation, including PTX.
-        '-keep' if keep_ptx else '',
-        # the target PTX and SASS version.
-        '-gencode',
-        f'arch=compute_{cc_code},code=sm_{cc_code}',
-        # allow ptxas (PTX assembler) to output information like register/smem usage.
-        '--ptxas-options=-v',
-        # compile into position independent code.
-        '--compiler-options',
-        "'-fPIC'",
-        # embed the line information into the binary, allow Nsight Compute to get the source code for profiling.
-        '-lineinfo',
-        # link the hidet runtime, all APIs for communication between kernels and host system are in hidet runtime.
-        '-lhidet_runtime',
-        # shared cuda runtime library is used (.so), instead of static one (.a). used to reduce binary size.
-        '--cudart',
-        'shared',
-        # supress some warnings
-        # see https://docs.nvidia.com/cuda/cuda-compiler-driver-nvcc/index.html#generic-tool-options-diag-suppress
-        # supress warming no 177 like: "warning #177-D: variable "xxx" was declared but never referenced"
-        '--diag-suppress 177',
-        # supress warning no 179 like: "warning #179-D: right operand of "%" is zero"
-        '--diag-suppress 179',
-        # supress warning no 39 like: "warning #39-D: division by zero"
-        '--diag-suppress 39',
-        # generate shared library (lib.so).
-        '--shared',
-        # the source path.
-        src_path,
-        # the output library path.
-        '-o',
-        out_lib_path,
-    ]
+    command = compile_command(src_path, out_lib_path, target, keep_ptx)
 
     try:
         # the directory to store the library "lib.so"
@@ -138,19 +189,21 @@ def compile_source(src_path: str, out_lib_path: str, keep_ptx=False) -> None:
                     message += result.stdout.decode().strip() + '\n'
                 if result.stderr:
                     message += result.stderr.decode().strip()
-                ptx_name = os.path.basename(src_path).replace('.cu', '.ptx')
-                if keep_ptx and os.path.exists(os.path.join(working_dir, ptx_name)):
+                if target == 'cuda':
+                    ptx_name = os.path.basename(src_path).replace('.cu', '.ptx')
+                    if keep_ptx and os.path.exists(os.path.join(working_dir, ptx_name)):
+                        ptx_path = os.path.join(working_dir, ptx_name)
+                        target_ptx_path = os.path.join(out_lib_dir, ptx_name)
+                        shutil.move(ptx_path, target_ptx_path)
+                raise CompilationFailed(src_path, message)
+            if target == 'cuda':
+                if keep_ptx:
+                    ptx_name = os.path.basename(src_path).replace('.cu', '.ptx')
                     ptx_path = os.path.join(working_dir, ptx_name)
                     target_ptx_path = os.path.join(out_lib_dir, ptx_name)
                     shutil.move(ptx_path, target_ptx_path)
-                raise CompilationFailed(src_path, message)
-            if keep_ptx:
-                ptx_name = os.path.basename(src_path).replace('.cu', '.ptx')
-                ptx_path = os.path.join(working_dir, ptx_name)
-                target_ptx_path = os.path.join(out_lib_dir, ptx_name)
-                shutil.move(ptx_path, target_ptx_path)
                 # os.rename(ptx_path, target_ptx_path)
-            with open(os.path.join(out_lib_dir, 'nvcc_log.txt'), 'w') as f:
+            with open(os.path.join(out_lib_dir, log_name(target) + '.txt'), 'w') as f:
                 output = '\n'.join([result.stdout.decode('utf-8').strip(), result.stderr.decode('utf-8').strip()])
                 f.write(output)
